Query processing test: added test of delete statement with where clause. Removed console write in a test.

+++ src/Samples/QueryProcessingTest/DataPopulation.cs
@@ -48,7 +48,7 @@ namespace QueryProcessingTest {
 
         public static void DeleteAccounts() {
             Db.Transaction(delegate {
-                Db.SlowSQL("DELETE FROM Account");
+                Db.SlowSQL("DELETE FROM Account where Accountid >= 0");
                 Db.SlowSQL("DELETE FROM QueryProcessingTest.User");
             });
         }
